Print kubekeeper output to stderr
When printing to stdout, kubekeeper output is unexpectedly piped to the
following command, breaking the assumption that kubekeeper is a
transparent wrapper to kubectl. This behavior is even more apparent when
redirecting the output to a file, the command hangs waiting for the user
input, while nothing is displayed at all.